backend/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
amqp==5.3.1
anyio==4.9.0
billiard==4.2.1
blis==1.2.1
catalogue==2.0.10
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
commonmark==0.9.1
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dbml==0.1.0
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
kombu==5.5.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.12
numpy==2.2.4
openai==1.73.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.55
preshed==3.0.9
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==1.10.13
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==10.16.2
rsa==4.9
setuptools==78.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.4
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.1
stripe==12.0.0
tenacity==9.1.2
thinc==8.3.4
tiktoken==0.9.0
tldextract==5.2.0
tqdm==4.67.1
typer==0.4.2
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
python-multipart
mailersend
better_profanity
werkzeug